Add an event hub Vue component to handle resetting podcasts when toggling between them.
This commit solves the problem where toggling to another podcast would still leave the previous one running. Both Podcast and PodcastContainer components subscribe to the event hub component, which is just an empty Vue instance. The Podcast component gets a 'reset' method that stops it from playing (and, optionally, resets the current time back to 0 - currently disabled). It listens for 'reset' events on the event hub. The PodcastContainer emits 'reset' events whenever someone clicks on the 'Next' or 'Previous' buttons.